+++ csrc/hooks.c
@@ -24,7 +24,7 @@
 #include <string.h>
 #include <unicorn/unicorn.h>
 
-/// @submodule unicorn_c_
+#define UL_IS_HOOKING_CPUID_SUPPORTED (UC_VERSION_MAJOR >= 2)
 
 static void get_common_arguments(lua_State *restrict L, ULHookState *restrict hook,
                                  uc_engine **engine, uc_hook_type *restrict hook_type,
@@ -39,23 +39,35 @@ static bool ulinternal_hook_callback__invalid_mem_access(uc_engine *engine,
 static uint32_t ulinternal_hook_callback__port_in(uc_engine *engine, uint32_t port,
                                                   int size, void *userdata);
 
+#if UL_IS_HOOKING_CPUID_SUPPORTED
+static bool ulinternal_hook_callback__cpuid(uc_engine *engine, void *userdata);
+#endif
+
 /* ISO C forbids casting a function pointer to an object pointer (void* in this case). As
  * Unicorn requires us to do this, we have to disable pedantic warnings temporarily so
  * that the compiler doesn't blow up. */
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wpedantic"
 
-int ul_create_arm64_sys_hook(lua_State *L)
+UL_NORETURN_MARKER int ul_create_arm64_sys_hook(lua_State *L)
 {
     ulinternal_crash_not_implemented(L);
 }
 
 int ul_create_cpuid_hook(lua_State *L)
 {
-    ulinternal_crash_not_implemented(L);
+    // Hooking the CPUID instruction is only supported on Unicorn 2.0+.
+#if UL_IS_HOOKING_CPUID_SUPPORTED
+    ulinternal_helper_create_code_hook(L, "cpuid",
+                                       (void *)ulinternal_hook_callback__cpuid);
+    return 1;
+#else
+    ulinternal_crash(
+        L, "Hooking the CPUID instruction isn't supported in your version of Unicorn.");
+#endif
 }
 
-int ul_create_edge_generated_hook(lua_State *L)
+UL_NORETURN_MARKER int ul_create_edge_generated_hook(lua_State *L)
 {
 #ifndef UC_HOOK_EDGE_GENERATED
     ulinternal_crash_unsupported_operation(L);
@@ -98,7 +110,7 @@ int ul_create_port_in_hook(lua_State *L)
     return 1;
 }
 
-int ul_create_tcg_opcode_hook(lua_State *L)
+UL_NORETURN_MARKER int ul_create_tcg_opcode_hook(lua_State *L)
 {
 #ifndef UC_HOOK_TCG_OPCODE
     ulinternal_crash_unsupported_operation(L);
@@ -109,6 +121,18 @@ int ul_create_tcg_opcode_hook(lua_State *L)
 
 #pragma GCC diagnostic pop
 
+/**
+ * Get the values of arguments passed to all hook creation functions.
+ *
+ * @param L The Lua state.
+ * @param[out] hook A pointer to an uninitialized @ref ULHookState struct.
+ * @param[out] engine A pointer that will be initialized to the `uc_engine *`
+ * @param[out] hook_type
+ * @param[out] start_address A pointer to a uint64_t that will be set to the lowest
+ *  address the hook is effective for.
+ * @param[out] end_address A pointer to a uint64_t that will be set to the highest address
+ *  the hook is effective for.
+ */
 static void get_common_arguments(lua_State *restrict L, ULHookState *restrict hook,
                                  uc_engine **engine, uc_hook_type *restrict hook_type,
                                  uint64_t *restrict start_address,
@@ -130,6 +154,19 @@ static void get_common_arguments(lua_State *restrict L, ULHookState *restrict ho
     lua_settable(L, LUA_REGISTRYINDEX);
 }
 
+/**
+ * Create a non-instruction hook.
+ *
+ * The Lua stack must contain the five required arguments for creating a general hook at
+ * indexes 1-5. Use @ref ulinternal_helper_create_code_hook for creating a hook triggered
+ * when a certain instruction is executed.
+ *
+ * @param L The Lua state.
+ * @param human_readable A human-readable slug indicating what the hook is for.
+ * @param callback The callback function Unicorn will invoke when the hook is triggered.
+ *
+ * @see ulinternal_helper_create_code_hook
+ */
 void ulinternal_helper_create_generic_hook(lua_State *L, const char *human_readable,
                                            void *callback)
 {
@@ -153,6 +190,46 @@ void ulinternal_helper_create_generic_hook(lua_State *L, const char *human_reada
         "Failed to create hook of type %ld (called as `%s`) from address 0x%08" PRIX64
         " through 0x%08" PRIX64 " (start > end means \"all of memory\")",
         (long)hook_type, human_readable, start_address, end_address);
+}
+
+/**
+ * Create a generic instruction hook based on arguments passed in on the Lua stack.
+ *
+ * The Lua stack must contain the six required arguments for creating a code hook, at
+ * indexes 1-6.
+ *
+ * @param L The Lua state.
+ * @param human_readable A short slug indicating what the code hook is for.
+ * @param callback The callback function Unicorn will invoke when the hook is triggered.
+ *
+ * @see ulinternal_helper_create_generic_hook
+ */
+void ulinternal_helper_create_code_hook(lua_State *L, const char *human_readable,
+                                        void *callback)
+{
+    uc_engine *engine;
+    uint64_t start_address, end_address;
+    uc_hook_type hook_type;
+
+#if LUA_VERSION_NUM >= 504
+    ULHookState *hook_state = (ULHookState *)lua_newuserdatauv(L, sizeof(*hook_state), 0);
+#else
+    ULHookState *hook_state = (ULHookState *)lua_newuserdata(L, sizeof(*hook_state));
+#endif
+    get_common_arguments(L, hook_state, &engine, &hook_type, &start_address,
+                         &end_address);
+
+    int instruction_id = (int)luaL_checkinteger(L, 6);
+
+    uc_err error = uc_hook_add(engine, &hook_state->hook_handle, hook_type, callback,
+                               hook_state, start_address, end_address, instruction_id);
+
+    ulinternal_crash_if_failed(
+        L, error,
+        "Failed to create hook of type %ld (called as `%s`) from address 0x%08" PRIX64
+        " through 0x%08" PRIX64
+        " (start > end means \"all of memory\") for instruction %d.",
+        (long)hook_type, human_readable, start_address, end_address, instruction_id);
 }
 
 int ul_hook_del(lua_State *L)
@@ -198,6 +275,14 @@ int ul_hook_del(lua_State *L)
     return 0;
 }
 
+/**
+ * Remove strong references for one or more callback functions used by hooks.
+ *
+ * Engine destructors use this to efficiently remove all hook callbacks at once.
+ *
+ * @param L The Lua state to get the callbacks for.
+ * @return This returns to Lua the number of callbacks that were possibly deallocated.
+ */
 int ul_release_hook_callbacks(lua_State *L)
 {
     int total_arguments = lua_gettop(L);
@@ -215,6 +300,11 @@ int ul_release_hook_callbacks(lua_State *L)
     return 1;
 }
 
+/**
+ * Push the Lua function handler for the given hook onto the Lua stack.
+ *
+ * @param hook The hook to get the callback function for.
+ */
 void ulinternal_push_callback_to_lua(const ULHookState *hook)
 {
     /* Retrieve the callback from the registry using this hook's metadata as the key. */
@@ -231,6 +321,19 @@ void ulinternal_push_callback_to_lua(const ULHookState *hook)
     }
 }
 
+/**
+ * A wrapper that will call the user's invalid memory access hook.
+ *
+ * @param engine The Unicorn engine this hook was triggered by.
+ * @param type The type of memory access.
+ * @param address The address an invalid access was attempted on.
+ * @param size The size of the memory operation.
+ * @param value For writes, the value attempted to write to memory.
+ * @param userdata A pointer to the @ref ULHookState for this hook.
+ *
+ * @return A boolean indicating whether this hook handled the exception (true) or whether
+ * to fall back to the default behavior (false).
+ */
 static bool ulinternal_hook_callback__invalid_mem_access(uc_engine *engine,
                                                          uc_mem_type type,
                                                          uint64_t address, int size,
@@ -263,6 +366,16 @@ static bool ulinternal_hook_callback__invalid_mem_access(uc_engine *engine,
     return return_value != 0;
 }
 
+/**
+ * A wrapper that will call the user's port read instruction hook.
+ *
+ * @param engine The Unicorn engine that called the hook.
+ * @param port The number of the port the instruction attempted to read from.
+ * @param size The size in bytes of the attempted port read (1, 2, or 4).
+ * @param userdata A pointer to the @ref ULHookState for this hook.
+ *
+ * @return The simulated return value from the port read.
+ */
 static uint32_t ulinternal_hook_callback__port_in(uc_engine *engine, uint32_t port,
                                                   int size, void *userdata)
 {
@@ -278,3 +391,35 @@ static uint32_t ulinternal_hook_callback__port_in(uc_engine *engine, uint32_t po
     lua_pop(hook->L, 1);
     return return_value;
 }
+
+#if UL_IS_HOOKING_CPUID_SUPPORTED
+/**
+ *
+ * @param engine The Unicorn engine invoking the hook.
+ * @param userdata A pointer to the @ref ULHookState for this hook.
+ * @return A boolean indicating whether the hook handled the instruction (true) or if
+ *  Unicorn should fall back to the default behavior (false).
+ */
+static bool ulinternal_hook_callback__cpuid(uc_engine *engine, void *userdata)
+{
+    (void)engine;
+    ULHookState *hook = (ULHookState *)userdata;
+
+    ulinternal_push_callback_to_lua(hook);
+    lua_call(hook->L, 0, 1);
+
+    // The function *must* return a boolean.
+    if (lua_type(hook->L, -1) != LUA_TBOOLEAN)
+    {
+        luaL_error(hook->L,
+                   "Error: The handler for the CPUID instruction must return a"
+                   " boolean telling Unicorn whether to skip the default behavior (true)"
+                   " or not (false).");
+        UL_UNREACHABLE_MARKER;
+    }
+
+    int return_value = lua_toboolean(hook->L, -1);
+    lua_pop(hook->L, 1);
+    return return_value != 0;
+}
+#endif