+++ tests/lua/hooks_spec.lua
@@ -204,4 +204,137 @@ describe('Hook tests', function ()
     assert.are.equals(0xf00d, uc:reg_read(info[1]), 'Register not written to')
     uc:close()
   end)
+
+  describe('[x86] CPUID', function()
+    it('Intercepted  #unicorn2only', function ()
+      local uc = unicorn.open(uc_const.UC_ARCH_X86, uc_const.UC_MODE_32)
+
+      local callback = function (...)
+        local argv = {...}
+        -- The only arguments we should be getting are the engine and the userdata.
+        assert_argument_count(argv, 2)
+        assert.are.equals(uc, argv[1])
+        assert.are.equals("stuff", argv[2])
+        uc:reg_write(x86.UC_X86_REG_EAX, 0x80000001)
+        uc:reg_write(x86.UC_X86_REG_EBX, 0)
+        uc:reg_write(x86.UC_X86_REG_ECX, 0)
+        uc:reg_write(x86.UC_X86_REG_EDX, 0)
+
+        -- Tell Unicorn we did something.
+        return true
+      end
+
+      uc:mem_map(0, 2^20)
+
+      -- xor eax, eax
+      -- cpuid
+      -- Opcodes: 31 c0 0f a2
+      uc:mem_write(0x7c000, '\049\192\015\162')
+      local handle = uc:hook_add(uc_const.UC_HOOK_INSN, callback, 0, 2^20, "stuff",
+        x86.UC_X86_INS_CPUID)
+      assert.not_nil(handle)
+
+      uc:emu_start(0x7c000, 0, 0, 2)
+      uc:emu_stop()
+      assert.are.equals(0x80000001, uc:reg_read(x86.UC_X86_REG_EAX))
+      assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_EBX))
+      assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_ECX))
+      assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_EDX))
+    end)
+
+    it('Not intercepted  #unicorn2only', function ()
+      local uc = unicorn.open(uc_const.UC_ARCH_X86, uc_const.UC_MODE_64)
+      uc:ctl_set_cpu_model(x86.UC_CPU_X86_HASWELL)
+
+      local callback = function (...)
+        local argv = {...}
+        -- The only arguments we should be getting are the engine and the userdata.
+        assert_argument_count(argv, 2)
+        assert.are.equals(uc, argv[1])
+        assert.are.equals("stuff", argv[2])
+        -- Don't do anything, tell Unicorn to resume the default behavior.
+        return false
+      end
+
+      uc:mem_map(0, 2^20)
+      uc:reg_write(x86.UC_X86_REG_EAX, 0)
+      uc:reg_write(x86.UC_X86_REG_EBX, 0)
+      uc:reg_write(x86.UC_X86_REG_ECX, 0)
+      uc:reg_write(x86.UC_X86_REG_EDX, 0)
+
+      -- cpuid = 0f a2
+      uc:mem_write(0x7c000, '\015\162')
+      local handle = uc:hook_add(uc_const.UC_HOOK_INSN, callback, 0, 2^20, "stuff",
+        x86.UC_X86_INS_CPUID)
+      assert.not_nil(handle)
+
+      uc:emu_start(0x7c000, 0, 0, 2)
+      uc:emu_stop()
+
+      -- CPUID is broken on Unicorn 2.1.3 or something. It's driving me nuts
+      -- that tests pass in CI on 2.1.4 but fail locally, so I'm supplying
+      -- both cases here.
+      if uc:reg_read(x86.UC_X86_REG_EAX) == 0 then
+        -- Unicorn 2.1.3
+        assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_EAX))
+        assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_EBX))
+        assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_ECX))
+        assert.are.equals(0, uc:reg_read(x86.UC_X86_REG_EDX))
+      else
+        -- EBX, EDX, and ECX (in that order) contain "GenuineIntel" in ASCII,
+        -- with characters ordered in little endian (BL = 'G', BH = 'e', etc.)
+        --    EBX = 'uneG'
+        --    EDX = 'Ieni'
+        --    ECX = 'letn
+        assert.are.equals(13, uc:reg_read(x86.UC_X86_REG_EAX))
+        assert.are.equals(0x756e6547, uc:reg_read(x86.UC_X86_REG_EBX))
+        assert.are.equals(0x49656e69, uc:reg_read(x86.UC_X86_REG_EDX))
+        assert.are.equals(0x6c65746e, uc:reg_read(x86.UC_X86_REG_ECX))
+      end
+
+      uc:hook_del(handle)
+    end)
+
+    it('Returning non-bool crashes  #unicorn2only #noluajit', function()
+      local uc = unicorn.open(uc_const.UC_ARCH_X86, uc_const.UC_MODE_32)
+
+      uc:mem_map(0, 2^20)
+      uc:mem_write(0x7c000, '\049\192\015\162')
+      local handle = uc:hook_add(
+        uc_const.UC_HOOK_INSN,
+        function (...) return 0 end,
+        0,
+        2^20,
+        "stuff",
+        x86.UC_X86_INS_CPUID
+      )
+
+      assert.has_error(
+        function () uc:emu_start(0x7c000, 0, 0, 2) end,
+        'Error: The handler for the CPUID instruction must return a boolean' ..
+        ' telling Unicorn whether to skip the default behavior (true) or not' ..
+        ' (false).'
+      )
+      uc:emu_stop()
+      uc:hook_del(handle)
+    end)
+
+    it('[x86] Hooking CPUID on Unicorn 1.x fails  #unicorn1only', function ()
+      local uc = unicorn.open(uc_const.UC_ARCH_X86, uc_const.UC_MODE_32)
+
+      uc:mem_map(0, 2^20)
+      uc:mem_write(0x7c000, '\049\192\015\162')
+      function try_to_create_hook()
+        local callback = function (...)
+          error("This should never have been called")
+        end
+        uc:hook_add(uc_const.UC_HOOK_INSN, callback, 0, 2^20, "stuff", x86.UC_X86_INS_CPUID)
+      end
+
+      assert.has_error(
+        try_to_create_hook,
+        "Hooking the CPUID instruction isn't supported in your version of Unicorn."
+      )
+    end)
+  end)
 end)